# W6-miniproject/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Javascript Study",
    "description": "항해99 W6 Miniproject"
  },
  "host": "15.165.204.0",
  "basePath": "/api",
  "tags": [
    {
      "name": "Users",
      "description": "로그인, 회원가입, 내 정보 조회"
    },
    {
      "name": "Posts",
      "description": "게시글 생성, 조회, 수정, 삭제"
    },
    {
      "name": "Comments",
      "description": "댓글 생성, 조회, 수정, 삭제"
    }
  ],
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-KEY",
      "description": "any description..."
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/posts": {
      "post": {
        "tags": [
          "Posts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Posts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/posts/{postId}": {
      "get": {
        "tags": [
          "Posts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Posts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Posts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/myInfo": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/signup": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "nickname": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "confirmPassword": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/posts/{postId}/comments": {
      "get": {
        "tags": [
          "Comments"
        ],
        "description": "",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/posts/{postId}/comment": {
      "post": {
        "tags": [
          "Comments"
        ],
        "description": "",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "comment": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/posts/{postId}/comment/{commentId}": {
      "put": {
        "tags": [
          "Comments"
        ],
        "description": "",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "comment": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": [
          "Comments"
        ],
        "description": "",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "comment": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  },
  "definitions": {
    "Users": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "test@google.com"
        },
        "nickname": {
          "type": "string",
          "example": "nickname"
        },
        "password": {
          "type": "string",
          "example": "test1234"
        },
        "profileImg": {
          "type": "string",
          "example": "url"
        }
      }
    },
    "Posts": {
      "type": "object",
      "properties": {
        "postId": {
          "type": "number",
          "example": 1
        },
        "nickname": {
          "type": "string",
          "example": "nickname"
        },
        "category": {
          "type": "string",
          "example": "초급"
        },
        "date": {
          "type": "string",
          "example": "2022-06-13"
        },
        "title": {
          "type": "string",
          "example": "title test"
        },
        "content": {
          "type": "string",
          "example": "content test"
        }
      },
      "required": [
        "nickname"
      ]
    },
    "Comments": {
      "type": "object",
      "properties": {
        "postId": {
          "type": "number",
          "example": 1
        },
        "commentId": {
          "type": "number",
          "example": 1
        },
        "nickname": {
          "type": "string",
          "example": "nickname"
        },
        "comment": {
          "type": "string",
          "example": "comment test"
        },
        "date": {
          "type": "string",
          "example": "2022-06-13"
        }
      },
      "required": [
        "postId",
        "nickname"
      ]
    }
  }
}